у
у

+++ калькулятор.cs
@@ -16,23 +16,23 @@ namespace ConsoleApp9
             {
                 case "сложение":
                     Console.Write("Введите первое слагаемое: ");
-                    double sl1 = Convert.ToInt32(Console.ReadLine());
+                    double sl1 = Convert.ToDouble(Console.ReadLine());
                     Console.Write("Введите второе слагаемое: ");
-                    double sl2 = Convert.ToInt32(Console.ReadLine());
+                    double sl2 = Convert.ToDouble(Console.ReadLine());
                     Console.Write($"Сумма = {sl1+sl2}");
                     break;
                 case "вычитание":
                     Console.Write("Введите уменьшаемое: ");
-                    double v1 = Convert.ToInt32(Console.ReadLine());
+                    double v1 = Convert.ToDouble(Console.ReadLine());
                     Console.Write("Введите вычитаемое: ");
-                    double v2 = Convert.ToInt32(Console.ReadLine());
+                    double v2 = Convert.ToDouble(Console.ReadLine());
                     Console.Write($"Разность = {v1 - v2}");
                     break;
                 case "деление":
                     Console.Write("Введите делимое: ");
-                    double d1 = Convert.ToInt32(Console.ReadLine());
+                    double d1 = Convert.ToDouble(Console.ReadLine());
                     Console.Write("Введите делитель: ");
-                    double d2 = Convert.ToInt32(Console.ReadLine());
+                    double d2 = Convert.ToDouble(Console.ReadLine());
                     if (d2 == 0)
                     {
                         Console.WriteLine("На 0 делить нельзя");
@@ -44,28 +44,28 @@ namespace ConsoleApp9
                     break;
                 case "умножение":
                     Console.Write("Введите первый множитель: ");
-                    double m1 = Convert.ToInt32(Console.ReadLine());
+                    double m1 = Convert.ToDouble(Console.ReadLine());
                     Console.Write("Введите второй множитель: ");
-                    double m2 = Convert.ToInt32(Console.ReadLine());
+                    double m2 = Convert.ToDouble(Console.ReadLine());
                     Console.Write($"Произведение = {m1*m2}");
                     break;
                 case "степень":
                     Console.Write("Введите число: ");
-                    double ch = Convert.ToInt32(Console.ReadLine());
+                    double ch = Convert.ToDouble(Console.ReadLine());
                     Console.Write("Введите степень: ");
-                    double s = Convert.ToInt32(Console.ReadLine());
+                    double s = Convert.ToDouble(Console.ReadLine());
                     Console.Write($"{ch} в {s} степени = {Math.Pow(ch,s)}");
                     break;
                 case "корень":
                     Console.Write("Введите число: ");
-                    double k = Convert.ToInt32(Console.ReadLine());
+                    double k = Convert.ToDouble(Console.ReadLine());
                     Console.Write($"Корень из {k} = {Math.Sqrt(k)}");
                     break;
                 case "остаток":
                     Console.Write("Введите делимое: ");
-                    double del1 = Convert.ToInt32(Console.ReadLine());
+                    double del1 = Convert.ToDouble(Console.ReadLine());
                     Console.Write("Введите делитель: ");
-                    double del2 = Convert.ToInt32(Console.ReadLine());
+                    double del2 = Convert.ToDouble(Console.ReadLine());
                     if (del2 == 0)
                     {
                         Console.WriteLine("На 0 делить нельзя");
@@ -77,11 +77,11 @@ namespace ConsoleApp9
                     break;
                 case "квадратное уравнение":
                     Console.Write("Введите a: ");
-                    double d = Convert.ToInt32(Console.ReadLine());
+                    double d = Convert.ToDouble(Console.ReadLine());
                     Console.Write("Введите b: ");
-                    double b = Convert.ToInt32(Console.ReadLine());
+                    double b = Convert.ToDouble(Console.ReadLine());
                     Console.Write("Введите c: ");
-                    double c = Convert.ToInt32(Console.ReadLine());
+                    double c = Convert.ToDouble(Console.ReadLine());
                     double D = Math.Sqrt(Math.Pow(b, 2) - 4 * d * c);
                     if (D >= 0)
                     {
